+++ src/main/java/de/markusbordihn/easynpc/client/renderer/entity/HumanoidSlimRenderer.java
@@ -22,6 +22,8 @@ package de.markusbordihn.easynpc.client.renderer.entity;
 import java.util.EnumMap;
 import java.util.Map;
 
+import com.mojang.blaze3d.vertex.PoseStack;
+
 import net.minecraft.Util;
 import net.minecraft.client.model.HumanoidModel;
 import net.minecraft.client.model.PlayerModel;
@@ -65,6 +67,7 @@ public class HumanoidSlimRenderer extends MobRenderer<EasyNPCEntity, PlayerModel
     this.addLayer(new ElytraLayer<>(this, context.getModelSet()));
   }
 
+  @Override
   public ResourceLocation getTextureLocation(EasyNPCEntity entity) {
     switch (entity.getSkinType()) {
       case PLAYER_SKIN:
@@ -73,6 +76,16 @@ public class HumanoidSlimRenderer extends MobRenderer<EasyNPCEntity, PlayerModel
         return PlayerTextureManager.getOrCreateTextureWithDefault(entity, DEFAULT_TEXTURE);
       default:
         return TEXTURE_BY_VARIANT.getOrDefault(entity.getVariant(), DEFAULT_TEXTURE);
+    }
+  }
+
+  @Override
+  protected void scale(EasyNPCEntity entity, PoseStack poseStack, float unused) {
+    if (entity.isBaby()) {
+      poseStack.scale(entity.getScaleX() * 0.5f, entity.getScaleY() * 0.5f,
+          entity.getScaleZ() * 0.5f);
+    } else {
+      poseStack.scale(entity.getScaleX(), entity.getScaleY(), entity.getScaleZ());
     }
   }