+++ src/main/java/de/markusbordihn/easynpc/client/screen/configuration/dialog/DialogConfigurationScreen.java
@@ -30,6 +30,7 @@ import net.minecraftforge.api.distmarker.OnlyIn;
 import de.markusbordihn.easynpc.Constants;
 import de.markusbordihn.easynpc.client.screen.configuration.ConfigurationScreen;
 import de.markusbordihn.easynpc.menu.configuration.ConfigurationMenu;
+import de.markusbordihn.easynpc.menu.configuration.ConfigurationType;
 import de.markusbordihn.easynpc.network.NetworkHandler;
 
 @OnlyIn(Dist.CLIENT)
@@ -49,15 +50,15 @@ public class DialogConfigurationScreen<T extends ConfigurationMenu> extends Conf
 
     // Dialog Types
     this.basicDialogButton =
-        this.addRenderableWidget(new Button(this.contentLeftPos, this.buttonTopPos, 80, 20,
+        this.addRenderableWidget(new Button(this.buttonLeftPos, this.buttonTopPos, 80, 20,
             new TranslatableComponent(Constants.TEXT_CONFIG_PREFIX + "basic_dialog"), onPress -> {
-              NetworkHandler.openDialog(uuid, "BasicDialogConfiguration");
+              NetworkHandler.openConfiguration(uuid, ConfigurationType.BASIC_DIALOG);
             }));
     this.yesNoDialogButton =
-        this.addRenderableWidget(new Button(this.contentLeftPos + this.basicDialogButton.getWidth(),
+        this.addRenderableWidget(new Button(this.buttonLeftPos + this.basicDialogButton.getWidth(),
             this.buttonTopPos, 80, 20,
             new TranslatableComponent(Constants.TEXT_CONFIG_PREFIX + "yes_no_dialog"), onPress -> {
-              NetworkHandler.openDialog(uuid, "YesNoDialogConfiguration");
+              NetworkHandler.openConfiguration(uuid, ConfigurationType.YES_NO_DIALOG);
             }));
 
     // Default button stats